+++ src/main/java/org/realityforge/replicant/server/transport/ReplicantSession.java
@@ -1,0 +1,20 @@
+package org.realityforge.replicant.server.transport;
+
+/**
+ * Base class for sessions within replicant.
+ */
+public abstract class ReplicantSession
+  extends org.realityforge.ssf.SimpleSessionInfo
+{
+  private final PacketQueue _queue = new PacketQueue();
+
+  public ReplicantSession( @javax.annotation.Nonnull final String sessionID )
+  {
+    super( sessionID );
+  }
+
+  public final PacketQueue getQueue()
+  {
+    return _queue;
+  }
+}